Move Result struct into lib.rs
