+++ p2panda-rs/src/atomic/entry_encoded.rs
@@ -1,11 +1,11 @@
+use crate::Result;
 use crate::atomic::Entry;
-use crate::error::Result;
 
 #[derive(Clone, Debug)]
 pub struct EntryEncoded(String);
 
 impl EntryEncoded {
-    pub fn new(value: String) -> Result<Self> {
+    pub fn new(_value: String) -> Result<Self> {
         todo!();
     }
