+++ p2panda-rs/src/atomic/seq_num.rs
@@ -2,8 +2,8 @@ use anyhow::bail;
 use bamboo_rs_core::lipmaa;
 use thiserror::Error;
 
+use crate::Result;
 use crate::atomic::Validation;
-use crate::error::Result;
 
 /// Start counting entries from here.
 pub const FIRST_SEQ_NUM: u64 = 1;
