--- p2panda-rs/src/error.rs
@@ -1,2 +1,0 @@
-/// A specialized `Result` type for p2panda-rs.
-pub type Result<T> = anyhow::Result<T>;
